requirements.txt
asgiref==3.7.2
Django==4.2.3
django-debug-toolbar==4.1.0
django-hosts==5.2
django-multiselectfield==0.1.12
Pillow==10.0.0
six==1.16.0
sqlparse==0.4.4
